Fix noise not being passed into the optimizer
We estimate the noise of each chess match using a Bayesian pentanomial model.
If this noise estimate is not passed into the optimizer, it will simply assume that
the noise is the same across the parameter space. This is not true of course,
since the variance changes with Elo strength.
While this looks like a minor bug, this was actually weakening the search a bit.